NATIVE-141: Implement DataContainer
